requirements.txt
pika
Flask
python-dotenv
